Cleanup class and document methods
This commit documents the methods and cleans up the class
